+++ crypto/zuc/build.info
@@ -1,2 +1,2 @@
 LIBS=../../libcrypto
-SOURCE[../../libcrypto]=zuc.c zuc_spec.c zuc_standard.c
+SOURCE[../../libcrypto]=zuc.c zuc_spec.c
