+++ util/libcrypto.num
@@ -4843,3 +4843,5 @@ o2i_SM2CiphertextValue                  4684	1_1_0d	EXIST::FUNCTION:
 i2o_SM2CiphertextValue                  4685	1_1_0d	EXIST::FUNCTION:
 SM2_compute_message_digest              4686	1_1_0d	EXIST::FUNCTION:
 serpent_set_decrypt_key                 4687	1_1_0d	EXIST::FUNCTION:
+sms4_standard_encrypt                   4688	1_1_0d	EXIST::FUNCTION:
+sms4_standard_decrypt                   4689	1_1_0d	EXIST::FUNCTION:
